+++ context/app/static/js/components/workspaces/LaunchWorkspaceDialog/LaunchWorkspaceDialog.tsx
@@ -1,8 +1,9 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import Button from '@mui/material/Button';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
 import LoadingButton from '@mui/lab/LoadingButton';
+import { useEventCallback } from '@mui/material/utils';
 
 import SummaryPaper from 'js/shared-styles/sections/SectionPaper';
 import DialogModal from 'js/shared-styles/dialogs/DialogModal';
@@ -17,6 +18,8 @@ import {
 } from 'js/components/workspaces/LaunchWorkspaceDialog/hooks';
 import { WorkspacesEventContextProvider } from 'js/components/workspaces/contexts';
 import { WorkspacesEventCategories } from 'js/components/workspaces/types';
+import { MAX_NUM_CONCURRENT_WORKSPACES } from 'js/components/workspaces/constants';
+import { tooManyWorkspacesRunning } from 'js/components/workspaces/utils';
 import WorkspaceEnvironmentDescription from '../WorkspaceEnvironmentDescription';
 
 const formId = 'launch-workspace-form';
@@ -43,8 +46,8 @@ const text = {
 
 function LaunchWorkspaceDialog() {
   const {
-    runningWorkspaceName,
-    runningWorkspaceIsCurrentWorkpace,
+    currentWorkspaceIsRunning,
+    runningWorkspaces,
     control,
     handleSubmit,
     submit,
@@ -59,9 +62,8 @@ function LaunchWorkspaceDialog() {
   const { toastErrorLaunchWorkspace } = useWorkspaceToasts();
 
   const newWorkspaceLaunch = dialogType === 'LAUNCH_NEW_WORKSPACE';
-  const isAnotherWorkspaceRunning = runningWorkspaceName && !runningWorkspaceIsCurrentWorkpace;
 
-  const runningWorkspaceAlert = isAnotherWorkspaceRunning && (
+  const tooManyWorkspacesAlert = tooManyWorkspacesRunning(runningWorkspaces) && !currentWorkspaceIsRunning && (
     <Alert
       severity="warning"
       sx={{
@@ -71,25 +73,27 @@ function LaunchWorkspaceDialog() {
         alignItems: 'center',
       }}
     >
-      {runningWorkspaceName} is currently running. You can only run one workspace at a time. To launch this workspace,
-      jobs in the workspace {runningWorkspaceName} will be stopped. Make sure to save all progress before launching this
-      workspace.
+      You can only run {MAX_NUM_CONCURRENT_WORKSPACES} workspaces at a time. Please stop at least one of your active
+      workspace jobs before launching this workspace.
     </Alert>
   );
 
-  const onSubmit = useCallback(
-    ({ workspaceJobTypeId, workspaceResourceOptions }: LaunchWorkspaceFormTypes) => {
-      submit({ workspaceJobTypeId, workspaceResourceOptions })
-        .then(() => {
-          handleClose();
-        })
-        .catch((e) => {
-          toastErrorLaunchWorkspace();
-          console.error(e);
-        });
-    },
-    [submit, handleClose, toastErrorLaunchWorkspace],
-  );
+  const onSubmit = useEventCallback(({ workspaceJobTypeId, workspaceResourceOptions }: LaunchWorkspaceFormTypes) => {
+    if (!workspace) {
+      toastErrorLaunchWorkspace();
+      console.error('No workspace to run found.');
+      return;
+    }
+
+    submit({ workspaceToLaunch: workspace, workspaceJobTypeId, workspaceResourceOptions })
+      .then(() => {
+        handleClose();
+      })
+      .catch((e) => {
+        toastErrorLaunchWorkspace();
+        console.error(e);
+      });
+  });
 
   return (
     <WorkspacesEventContextProvider
@@ -104,10 +108,10 @@ function LaunchWorkspaceDialog() {
           // eslint-disable-next-line @typescript-eslint/no-misused-promises
           <form id={formId} onSubmit={handleSubmit(onSubmit)}>
             {newWorkspaceLaunch ? (
-              runningWorkspaceAlert
+              tooManyWorkspacesAlert
             ) : (
               <Stack direction="column" spacing={1}>
-                {runningWorkspaceAlert}
+                {tooManyWorkspacesAlert}
                 <Alert severity="info">{text.resources.alert}</Alert>
                 <SummaryPaper>
                   <Stack direction="column" spacing={2}>
@@ -128,7 +132,13 @@ function LaunchWorkspaceDialog() {
             <Button type="button" onClick={handleClose} disabled={isSubmitting}>
               Cancel
             </Button>
-            <LoadingButton type="submit" variant="contained" form={formId} loading={isSubmitting}>
+            <LoadingButton
+              type="submit"
+              variant="contained"
+              disabled={tooManyWorkspacesRunning(runningWorkspaces)}
+              form={formId}
+              loading={isSubmitting}
+            >
               Launch Workspace
             </LoadingButton>
           </Stack>